// ProcMgt/ClientApp/src/app/models/model.ts
export class Model {
  modelID: string;
  modelCode: string;
  itemTypeID: string;
  makeID: string;
  modelName: string;
  isActive: boolean;
  image: string;
  uploadFileID: string;
  userID: string;
  entryDateTime: string;

  itemTypeName: string;
  makeName: string;
  uploadFileName: string;
  

}