// ProcMgt/ClientApp/src/app/models/purchaserequest.ts
export class Purchaserequest {


  itemRequestID: string;
  requestTitle: string;
  categoryID: string;
  itemID: string;
  itemTypeID: string;
  makeID: string;
  modelID: string;
  priorityID: string;
  requiredDate: string;
  isReplaceble: boolean;
  isIssued: boolean;
  assetCode: string;
  noOfUnits: number;
  approverID: string;
  remark: string;
  attachment: string; 
  isApproved: boolean;
  isRejected: boolean;
  issuedUserID: string;
  balancedQty: number;
  receivedQty: number;
  
  approvedDateTime: string;
  approvalComment: string;
  requestedDateTime: string;
  requestedUserID: string;
  departmentID: string;
  updatedRequestedDateTime: string;
  approvedUserID: string;

  categoryName: string;
  itemDescription: string;
  makeName: string;
  priorityLevelName: string;
  itemTypeName: string;
  modelName: string;
  approverName: string;
  requestedUserName: string;
  approvedUserName: string;
  departmentName: string;
  userID: string; //Add by Nipuna Francisku

}